move healthCheck to grpc client
