+++ internal/wrappers/grpcs/client.go
@@ -2,10 +2,13 @@ package grpcs
 
 import (
 	"context"
+	"fmt"
 	"time"
 
+	"github.com/checkmarx/ast-cli/internal/logger"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
+	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 )
 
 const (
@@ -14,12 +17,21 @@ const (
 
 type Client interface {
 	CreateClientConn() (*grpc.ClientConn, error)
-	GetContext() context.Context
+	HealthCheck(serviceName string) error
 }
 
 type BaseClient struct {
 	hostAddress string
 	ctx         context.Context
+}
+
+type ClientWithTimeout struct {
+	BaseClient
+	timeout time.Duration
+}
+
+func NewGRPCClientWithTimeout(hostAddress string, timeout time.Duration) Client {
+	return &ClientWithTimeout{BaseClient: BaseClient{hostAddress: hostAddress, ctx: context.Background()}, timeout: timeout}
 }
 
 func dialOptions() []grpc.DialOption {
@@ -33,13 +45,30 @@ func (c *BaseClient) dialContext(ctx context.Context) (*grpc.ClientConn, error) 
 	return grpc.DialContext(ctx, c.hostAddress, dialOptions()...)
 }
 
-type ClientWithTimeout struct {
-	BaseClient
-	timeout time.Duration
-}
+func (c *BaseClient) HealthCheck(serviceName string) error {
+	conn, connErr := c.CreateClientConn()
+	if connErr != nil {
+		return connErr
+	}
+	defer func(conn *grpc.ClientConn) {
+		_ = conn.Close()
+	}(conn)
 
-func NewGRPCClientWithTimeout(hostAddress string, timeout time.Duration) Client {
-	return &ClientWithTimeout{BaseClient: BaseClient{hostAddress: hostAddress, ctx: context.Background()}, timeout: timeout}
+	req := &healthpb.HealthCheckRequest{
+		Service: serviceName,
+	}
+
+	healthRes, healthErr := healthpb.NewHealthClient(conn).Check(c.ctx, req)
+	if healthErr != nil {
+		logger.PrintIfVerbose(fmt.Sprintf("Health Check Failed: %v", healthErr))
+		return healthErr
+	}
+
+	if healthRes.Status == healthpb.HealthCheckResponse_SERVING {
+		return nil
+	}
+
+	return fmt.Errorf("service not serving, status: %v", healthRes.Status)
 }
 
 func (c *ClientWithTimeout) CreateClientConn() (*grpc.ClientConn, error) {
@@ -48,6 +77,6 @@ func (c *ClientWithTimeout) CreateClientConn() (*grpc.ClientConn, error) {
 	return c.dialContext(ctx)
 }
 
-func (c *ClientWithTimeout) GetContext() context.Context {
-	return c.ctx
+func (c *BaseClient) CreateClientConn() (*grpc.ClientConn, error) {
+	return c.dialContext(c.ctx)
 }
